Assignment 3: Preparing setMatches
setMatches will be a function that takes auto-matched volunteers, categorizes them by event, and lets admins match or reject them from an event. Upon doing either or pressing the refresh button in the HTML file, the function will run again, clearing the list of events and volunteers and filling them with an updated list.